Replace IgnoredPathFileSnapshot by FileContentSnapshot
