+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/state/FileHashSnapshot.java
@@ -20,7 +20,7 @@ import com.google.common.base.Objects;
 import org.gradle.internal.file.FileType;
 import org.gradle.internal.hash.HashCode;
 
-public class FileHashSnapshot implements FileContentSnapshot {
+public class FileHashSnapshot extends AbstractFileContentSnapshot {
     private final HashCode hash;
     private final transient long lastModified; // Currently not persisted
